+++ src/enum/ActionEnum.ts
@@ -1,0 +1,4 @@
+export enum ActionEnum {
+  Add = 'Add',
+  Remove = 'Remove',
+}